Runnable unittest reorganization
There are some "blind spots" in the runnable code, such as when
dealing with serialization from JSON files and Runnable's
configuration.

To allow for a clearer view of the current coverage of Runnable's
features, let's reorganize their tests in a file of its own, and based
on the usage and where the data is coming from or going to.
